Restore CommonJS compatibility
