+++ test/integration/access.mjs
@@ -2,7 +2,7 @@ import assert from "assert";
 import request from "request";
 import { serve } from "../../dist/express-restify-mongoose.js";
 
-import setupDb from "./setup.js";
+import setupDb from "./setup.mjs";
 
 export default function (createFn, setup, dismantle) {
   const db = setupDb();
