+++ test/integration/options.mjs
@@ -3,7 +3,7 @@ import request from "request";
 import sinon from "sinon";
 import { serve } from "../../dist/express-restify-mongoose.js";
 
-import setupDb from "./setup.js";
+import setupDb from "./setup.mjs";
 
 export default function (createFn, setup, dismantle) {
   const db = setupDb();
