--- test/unit.js
@@ -1,12 +1,0 @@
-import "./unit/buildQuery.js";
-import "./unit/detective.js";
-import "./unit/errorHandler.js";
-import "./unit/middleware/access.js";
-import "./unit/middleware/ensureContentType.js";
-import "./unit/middleware/onError.js";
-import "./unit/middleware/outputFn.js";
-import "./unit/middleware/prepareOutput.js";
-import "./unit/middleware/prepareQuery.js";
-import "./unit/moredots.js";
-import "./unit/resourceFilter.js";
-import "./unit/weedout.js";
