+++ test/unit.mjs
@@ -1,0 +1,12 @@
+import "./unit/buildQuery.mjs";
+import "./unit/detective.mjs";
+import "./unit/errorHandler.mjs";
+import "./unit/middleware/access.mjs";
+import "./unit/middleware/ensureContentType.mjs";
+import "./unit/middleware/onError.mjs";
+import "./unit/middleware/outputFn.mjs";
+import "./unit/middleware/prepareOutput.mjs";
+import "./unit/middleware/prepareQuery.mjs";
+import "./unit/moredots.mjs";
+import "./unit/resourceFilter.mjs";
+import "./unit/weedout.mjs";
