+++ test/unit/middleware/prepareOutput.mjs
